pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "syllabus"
version = "0.1.0"
description = "AI fertility learning platform - pipeline and CLI (V0)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic[email]>=2.0",
    "typer[all]>=0.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "httpx>=0.27",
    "chromadb>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=4.0",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["syllabus*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["syllabus/tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100
exclude = [".git", "__pycache__", ".venv", "venv", "build", "dist"]

[tool.ruff.lint]
select = ["F", "E", "I", "B", "C4"]
ignore = ["E501"]  # line length; formatter handles most; long prompts excluded

[tool.ruff.format]

[tool.coverage.run]
source = ["syllabus"]
branch = true

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "def __repr__", "if TYPE_CHECKING:"]
